+++ index.html
@@ -249,7 +249,7 @@
   </template>
 
   <!-- Подключение внешнего скрипта -->
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" defer type="module"></script>
 </body>
 
 </html>
